change url and password regex, fix add authorId on save message and back to live custom jwt validation

+++ api/src/messages/controllers/messages.controller.ts
@@ -5,7 +5,9 @@ import {
     Body,
     Patch,
     Delete,
-    UseGuards, Param
+    UseGuards,
+    Param,
+    ForbiddenException
 } from "@nestjs/common";
 import { MessagesService } from "messages/services/messages.service";
 import { CreateMessageDto } from "messages/model/dto/create-message.dto";
@@ -13,19 +15,29 @@ import { UpdateMessageDto } from "messages/model/dto/update-message.dto";
 import {
     ApiBadRequestResponse,
     ApiBearerAuth,
-    ApiCreatedResponse, ApiForbiddenResponse,
+    ApiCreatedResponse,
+    ApiForbiddenResponse,
     ApiNotFoundResponse,
     ApiOkResponse,
     ApiTags,
     ApiUnauthorizedResponse
 } from "@nestjs/swagger";
 import { JwtAuthGuard } from "auth/guards/jwt-auth.guard";
+import { ConfigService } from "@nestjs/config";
+import { UserReq } from "users/decorators/user.decorator";
 
-@ApiBearerAuth('defaultBearerAuth')
+@ApiBearerAuth("defaultBearerAuth")
 @ApiTags("messages")
 @Controller("messages")
 export class MessagesController {
-    constructor(private readonly messagesService: MessagesService) {}
+    private readonly isDevelopment: boolean;
+
+    constructor(
+        private readonly messagesService: MessagesService,
+        private readonly configService: ConfigService
+    ) {
+        this.isDevelopment = configService.get("DEVELOPMENT") === "true";
+    }
 
     @ApiUnauthorizedResponse()
     @ApiCreatedResponse()
@@ -33,8 +45,10 @@ export class MessagesController {
     @UseGuards(JwtAuthGuard)
     @Post()
     async create(
-        @Body() createMessageDto: CreateMessageDto
+        @Body() createMessageDto: CreateMessageDto,
+        @UserReq("id") userId: number
     ) {
+        createMessageDto.authorId = userId;
         return this.messagesService.create(createMessageDto);
     }
 
@@ -47,21 +61,37 @@ export class MessagesController {
     }
 
     @ApiUnauthorizedResponse()
+    @ApiForbiddenResponse()
     @ApiOkResponse()
     @ApiNotFoundResponse()
     @UseGuards(JwtAuthGuard)
     @Get(":id")
-    async findOne(@Param("id") id: number) {
+    async findOne(@Param("id") id: number, @UserReq("id") userId: number) {
+        const user = await this.messagesService.findOne(id).catch(() => {
+            throw new ForbiddenException();
+        });
+
+        if (!this.isDevelopment && user.author.id !== userId)
+            throw new ForbiddenException();
+
         return this.messagesService.findOne(id);
     }
 
     @ApiUnauthorizedResponse()
+    @ApiForbiddenResponse()
     @ApiOkResponse()
     @ApiNotFoundResponse()
     @UseGuards(JwtAuthGuard)
     @Delete(":id")
-    async delete(@Param("id") id: number) {
-        return this.messagesService.delete(id);
+    async delete(@Param("id") id: number, @UserReq("id") userId: number) {
+        const user = await this.messagesService.findOne(id).catch(() => {
+            throw new ForbiddenException();
+        });
+
+        if (!this.isDevelopment && user.author.id !== userId)
+            throw new ForbiddenException();
+
+        await this.messagesService.delete(id);
     }
 
     @ApiUnauthorizedResponse()
@@ -73,8 +103,20 @@ export class MessagesController {
     @Patch(":id")
     async update(
         @Body() updateMessageDto: UpdateMessageDto,
-        @Param("id") id: number
+        @Param("id") id: number,
+        @UserReq("id") userId: number
     ) {
-        return this.messagesService.update(id, updateMessageDto);
+        const message = await this.messagesService.findOne(id).catch(() => {
+            throw new ForbiddenException();
+        });
+
+        if (!this.isDevelopment && message.author.id !== userId)
+            throw new ForbiddenException();
+
+        await this.messagesService.update(
+            id,
+            updateMessageDto,
+            message.isImage
+        );
     }
 }
